requirements.txt
flask==2.0.1
torch==1.9.0
transformers==4.9.2
python-dotenv==0.19.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.0.2
gunicorn==20.1.0
whitenoise==5.3.0 